feat: Invoke translateNote via API.
