+++ src/models/page.model.ts
@@ -22,6 +22,7 @@ import { checkSlugReservedWords } from "~/lib/slug-reserved-words"
 import { getKeys, getStorage } from "~/lib/storage"
 import {
   ExpandedNote,
+  Language,
   NoteType,
   PagesSortTypes,
   PageVisibilityEnum,
@@ -419,6 +420,7 @@ export async function getPage<TRender extends boolean = false>(input: {
   noteId?: number
   handle?: string // In order to be compatible with old drafts
   disableAutofill?: boolean
+  translateTo?: Language
 }) {
   const mustLocal = input.slug?.startsWith("!local-") && !input.noteId
 
@@ -570,6 +572,7 @@ export async function getPage<TRender extends boolean = false>(input: {
       note: page,
       useStat: input.useStat,
       disableAutofill: input.disableAutofill,
+      translateTo: input.translateTo,
     })
   }
